refactor(core): add improved transction write item model
