+++ packages/core/public-api.ts
@@ -8,7 +8,7 @@ export * from './src/classes/connection/connection-options';
 // models
 export * from './src/classes/expression/condition';
 export * from './src/classes/expression/key-condition';
-export * from './src/classes/transaction/write-transaction';
+export * from './src/classes/transaction/write-transaction-old';
 export * from './src/classes/batch/write-batch';
 export * from './src/classes/batch/read-batch';
