+++ packages/core/src/classes/batch/batch.ts
@@ -1,5 +1,12 @@
-export abstract class Batch {
-  constructor() {}
+export abstract class Batch<T> {
+  protected _items: T[];
+  constructor() {
+    this._items = [];
+  }
 
   abstract add(item: any): this;
+
+  get items() {
+    return this._items;
+  }
 }
