+++ packages/core/src/classes/batch/write-batch.ts
@@ -18,13 +18,7 @@ export type WriteBatchItem<Entity, PrimaryKey> =
   | WriteBatchCreate<Entity>
   | WriteBatchDelete<Entity, PrimaryKey>;
 
-export class WriteBatch extends Batch {
-  protected _items: WriteBatchItem<any, any>[];
-  constructor() {
-    super();
-    this._items = [];
-  }
-
+export class WriteBatch extends Batch<WriteBatchItem<any, any>> {
   add(batchItems: WriteBatchItem<any, any>[]): this {
     this.items.push(...batchItems);
     return this;
@@ -50,9 +44,5 @@ export class WriteBatch extends Batch {
       },
     });
     return this;
-  }
-
-  get items() {
-    return this._items;
   }
 }
