+++ packages/core/src/classes/manager/__test__/transaction-manager.spec.ts
@@ -1,7 +1,7 @@
 import {UserPrimaryKey} from '../../../../__mocks__/user';
 import {createTestConnection, resetTestConnection} from '@typedorm/testing';
 import {Connection} from '../../connection/connection';
-import {WriteTransaction} from '../../transaction/write-transaction';
+import {WriteTransaction} from '../../transaction/write-transaction-old';
 import {TransactionManager} from '../transaction-manager';
 import {User} from '../../../../__mocks__/user';
 import {
