+++ packages/core/src/classes/transaction/write-transaction-old.ts
@@ -1,0 +1,118 @@
+import {DynamoDB} from 'aws-sdk';
+import {dropProp} from '../../helpers/drop-prop';
+import {Connection} from '../connection/connection';
+import {
+  isLazyTransactionWriteItemListLoader,
+  LazyTransactionWriteItemListLoader,
+} from '../transformer/is-lazy-transaction-write-item-list-loader';
+import {
+  Transaction,
+  WriteTransactionChainItem,
+  WriteTransactionCreate,
+} from './transaction-old';
+import {
+  isCreateTransaction,
+  isRemoveTransaction,
+  isUpdateTransaction,
+  isWriteTransactionItemList,
+} from './type-guards';
+
+export class WriteTransaction extends Transaction {
+  protected _items: (
+    | LazyTransactionWriteItemListLoader
+    | DynamoDB.DocumentClient.TransactWriteItem
+  )[];
+
+  constructor(
+    connection: Connection,
+    initialItems?: DynamoDB.DocumentClient.TransactWriteItemList
+  ) {
+    super(connection);
+
+    // initialize items if there are any
+    if (initialItems) {
+      this._items = [...initialItems];
+    }
+  }
+
+  /**
+   * @deprecated use operation specific method or `.add` instead
+   */
+  chian<PrimaryKey, Entity>(
+    chainedItem: WriteTransactionChainItem<PrimaryKey, Entity>
+  ): WriteTransaction {
+    // create
+    if (isCreateTransaction<Entity>(chainedItem)) {
+      this.items = this.chainCreateTransaction(chainedItem);
+      // update
+    } else if (isUpdateTransaction<PrimaryKey, Entity>(chainedItem)) {
+      const {item, body, primaryKey, options} = chainedItem.update;
+
+      const itemToUpdate = this._dcReqTransformer.toDynamoUpdateItem<
+        PrimaryKey,
+        Entity
+      >(item, primaryKey, body, options);
+      if (!isLazyTransactionWriteItemListLoader(itemToUpdate)) {
+        this.items.push({
+          Update: dropProp(itemToUpdate, 'ReturnValues') as DynamoDB.Update,
+        });
+      } else {
+        this.items.push(itemToUpdate);
+      }
+      // remove
+    } else if (isRemoveTransaction<PrimaryKey, Entity>(chainedItem)) {
+      const {item, primaryKey, options} = chainedItem.delete;
+
+      const itemToRemove = this._dcReqTransformer.toDynamoDeleteItem<
+        PrimaryKey,
+        Entity
+      >(item, primaryKey, options);
+      if (!isLazyTransactionWriteItemListLoader(itemToRemove)) {
+        this.items.push({
+          Delete: itemToRemove,
+        });
+      } else {
+        this.items.push(itemToRemove);
+      }
+    } else {
+      return this;
+    }
+    return this;
+  }
+
+  private chainCreateTransaction<Entity>(
+    chainedItem: WriteTransactionCreate<Entity>
+  ) {
+    const {
+      create: {item, options},
+    } = chainedItem;
+
+    const dynamoPutItemInput = this._dcReqTransformer.toDynamoPutItem<Entity>(
+      item,
+      options
+    );
+
+    if (!isWriteTransactionItemList(dynamoPutItemInput)) {
+      return [
+        {
+          Put: dynamoPutItemInput,
+        },
+      ] as DynamoDB.DocumentClient.TransactWriteItemList;
+    }
+
+    return [...dynamoPutItemInput];
+  }
+
+  get items() {
+    return this._items;
+  }
+
+  set items(
+    items: (
+      | DynamoDB.DocumentClient.TransactWriteItem
+      | LazyTransactionWriteItemListLoader
+    )[]
+  ) {
+    this._items = [...this.items, ...items];
+  }
+}
